cdx: let the mark say what to remark egress frames with
cdx_ehash.c has always read qosmark->dscp_mark_flag and
dscp_mark_value, writing the codepoint into the header manipulation that
already decrements TTL, so the remark costs no extra opcode. Nothing in
the tree had ever written them: NXP left that to an iptables target
OpenWrt never packaged, which is the last parity gap against the
original ASK.

cdx_ft_rule.qos widens from twelve bits to nineteen -- u16 to u32, since
a u16 would silently drop the codepoint -- and the operator keeps
thirteen bits of ct->mark instead of twenty.

Carrying it in a tc filter instead would have cost no bits, and was
rejected on what OpenWrt can express: act_pedit and act_csum are not in
its core scheduler package, only in a fifteen-module bundle, where ct
mark set is nftables and already in every image. The operator is also
already setting ct mark for the class queue and the policer, so the
remark rides a rule that has to exist anyway rather than splitting one
flow's QoS across a mark and a filter.

The remark is the one field with a flag, and needs one: every other
field spells "unspecified" as zero, but DSCP zero is CS0 -- a real
codepoint, and the one an operator remarks to for best effort -- so
"remark to zero" and "do not remark" cannot be the same value.

Widening the class also makes CDX_FT_QOS_EGRESS_MASK load-bearing
rather than merely correct: unmasked, a remark of CS7 would index 508
entries past the Tx path's 256-entry class table.

Rig: 40 of 40 bulk packets left at tos 0xb8 with the flows in hardware
and 40 of 40 at tos 0x0 with the same traffic in software.

+++ cdx/cdx_flowtable.h
@@ -30,7 +30,7 @@ void cdx_unregister_ft_setup_tc(void);
  * Deriving it twice from the same mark would still be two decodes to keep in
  * step; this is one. Unregistered, the software path expresses no opinion and
  * behaves as it did before there was a qdisc. */
-typedef u16 (*cdx_ft_qos_class_fn)(u32 mark);
+typedef u32 (*cdx_ft_qos_class_fn)(u32 mark);
 int cdx_register_ft_qos_class(cdx_ft_qos_class_fn fn);
 void cdx_unregister_ft_qos_class(void);
